+++ doc/whats-new.rst
@@ -25,8 +25,8 @@ New Features
 - Allow chunking for arrays with duplicated dimension names (:issue:`8759`, :pull:`9099`).
   By `Martin Raspaud <https://github.com/mraspaud>`_.
 
-- New ``DataTree`` methods :py:meth:`xarray.core.datatree.DataTree.to_dict_nested`
-  and :py:meth:`xarray.core.datatree.DataTree.from_dict_nested` to allow easy conversion between
+- New ``DataTree`` methods :py:meth:`xarray.core.datatree.DataTree.to_nested_dict`
+  and :py:meth:`xarray.core.datatree.DataTree.from_nested_dict` to allow easy conversion between
   dictionaries and xarray objects
   (:issue:`9074`, :pull:`9080`).
   By `Etienne Schalk <https://github.com/etienneschalk>`_